Moved everything to explore tab
