+++ frontend/src/lib/event-details/EventDetails.scss
