+++ frontend/src/lib/event-details/EventDetails.tsx
