+++ frontend/src/queries/nodes/DataTable/EventRowActions.tsx
@@ -5,7 +5,7 @@ import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { More } from 'lib/lemon-ui/LemonButton/More'
 import { copyToClipboard } from 'lib/utils/copyToClipboard'
 import { getCurrentTeamId } from 'lib/utils/getAppContext'
-import { createActionFromEvent } from 'scenes/activity/explore/createActionFromEvent'
+import { createActionFromEvent } from 'scenes/explore/events/createActionFromEvent'
 import { insightUrlForEvent } from 'scenes/insights/utils'
 import { sessionPlayerModalLogic } from 'scenes/session-recordings/player/modal/sessionPlayerModalLogic'
 import { teamLogic } from 'scenes/teamLogic'
