+++ frontend/src/scenes/explore/events/EventsScene.tsx
