+++ frontend/src/scenes/explore/events/__mocks__/eventsQuery.json
