+++ frontend/src/scenes/explore/events/defaults.ts
