--- frontend/src/scenes/explore/explore/EventDetails.scss
@@ -1,3 +1,0 @@
-.LemonTabs[data-attr='event-details'] ul {
-    padding: 0 0.75rem;
-}
