+++ frontend/src/scenes/explore/live-events/LiveEventsTable.tsx
@@ -6,9 +6,10 @@ import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
 import { TZLabel } from 'lib/components/TZLabel'
 import { LemonTable, LemonTableColumns } from 'lib/lemon-ui/LemonTable'
-import { liveEventsTableLogic } from 'scenes/activity/live/liveEventsTableLogic'
 
 import type { LiveEvent } from '~/types'
+
+import { liveEventsTableLogic } from './liveEventsTableLogic'
 
 const columns: LemonTableColumns<LiveEvent> = [
     {
