+++ frontend/src/scenes/explore/live-events/liveEventsTableLogic.tsx
